requirements.txt
python-telegram-bot==12.7
httpcore==0.9.1
httpx==0.13.3
googletrans
gTTS
SpeechRecognition
ftransc
pymongo[tls,srv]
